Overview: Clean up. Details: 1. Deleted hash version of HopMap.go. 2. Improved BitArray.go. TODO: 1. Optimize linked list variant more. 2. Deal with empty element initialization problem. 3. Add the option to cache hash values for faster resize. 4. Delete linear version.

+++ BitArray.go
@@ -19,7 +19,7 @@ func (u BitArray) Len() int {
 }
 
 func (u BitArray) Get(i int) bool {
-	t := 1 << i & (bits.UintSize - 1)
+	t := uint(1 << (i & (bits.UintSize - 1)))
 	return u.bits[i>>LOG_UINT_SIZE]&t == t
 }
